#26 Extract Grammemes inner classes from English and Russian
